program for fun
This solution does amazingly well when there are only two inputs. The multinomial is a multi-variate generalization of the binomal probability. The example given in the documentation is a standard die rolled x number of times with the resulting array being the number of times the die landed on each of its faces.

The only problem is that since there are so many possible distributions with four inputs, the probability that the maximum I get after a certain number of iterations is so small that it takes multiple minutes to compute the wrong solution. 
Nonetheless, this was a fun solution that could consistently (and quickly) solve the test problem!